+++ packages/client/src/escrow/graphql/queries.ts
@@ -14,3 +14,27 @@ export const getProtocolAndPlatformsFees = (
       }
     }
   `;
+
+export const getPaymentsByService = (serviceId: string, paymentType?: string) => {
+  let condition = `where: {service: "${serviceId}"`;
+  paymentType ? (condition += `, paymentType: "${paymentType}"`) : '';
+  condition += '}, orderBy: id, orderDirection: asc';
+  const query = `
+    {
+      payments(${condition}) {
+        id
+        amount
+        rateToken {
+          address
+          decimals
+          name
+          symbol
+        }
+        paymentType
+        transactionHash
+        createdAt
+      }
+    }
+    `;
+  return query;
+};
