--- packages/client/src/fees/graphql/index.ts
@@ -1,16 +1,0 @@
-export const getProtocolAndPlatformsFees = (
-  originServicePlatformId: string,
-  originValidatedProposalPlatformId: string,
-) => `
-    {
-      protocols {
-        protocolEscrowFeeRate
-      }
-      servicePlatform: platform(id:${originServicePlatformId}){
-        originServiceFeeRate
-      }
-      proposalPlatform: platform(id:${originValidatedProposalPlatformId}){
-        originValidatedProposalFeeRate
-      }
-    }
-      `;
