--- packages/client/src/fees/types/index.ts
@@ -1,8 +1,0 @@
-export interface IFees {
-  getMintFees(): Promise<any>;
-
-  getProtocolAndPlatformsFees(
-    originServicePlatformId: string,
-    originValidatedProposalPlatformId: string,
-  ): Promise<any>;
-}
