+++ packages/client/src/profile/types/index.ts
@@ -14,6 +14,15 @@ export type TalentLayerProfile = {
 export interface IProfile {
   upload(profileData: TalentLayerProfile): Promise<string>;
   getByAddress(address: `0x${string}`): Promise<any>;
+  getById(userId: string): Promise<any>;
   create(handle: string): Promise<any>;
   update(profileData: TalentLayerProfile, userId: string): Promise<ClientTransactionResponse>;
+  getUsers(params: {
+    numberPerPage?: number;
+    offset?: number;
+    searchQuery?: string;
+  }): Promise<any>;
+  getTotalGains(userId: string): Promise<any>;
+  getPayments(userId: string, numberPerPage?: number, offset?: number, startDate?: string, endDate?: string): Promise<any>
+  getMintFees(): Promise<any>;
 }
