+++ packages/client/src/reviews/graphql/index.ts
@@ -1,0 +1,21 @@
+export const getReviewsByService = (serviceId: string) => `
+{
+  reviews(where: { service: "${serviceId}" }, orderBy: id, orderDirection: desc) {
+    id
+    rating
+    createdAt
+    service {
+      id
+      status
+    }
+    to {
+      id
+      handle
+    }
+    description{
+      id
+      content
+    }
+  }
+}
+`;
