+++ packages/client/src/services/graphql/queries.ts
@@ -1,51 +1,5 @@
 import { IProps } from '../types';
 
-export const getReviewsByService = (serviceId: string) => `
-{
-  reviews(where: { service: "${serviceId}" }, orderBy: id, orderDirection: desc) {
-    id
-    rating
-    createdAt
-    service {
-      id
-      status
-    }
-    to {
-      id
-      handle
-    }
-    description{
-      id
-      content
-    }
-  }
-}
-`;
-
-export const getPaymentsByService = (serviceId: string, paymentType?: string) => {
-  let condition = `where: {service: "${serviceId}"`;
-  paymentType ? (condition += `, paymentType: "${paymentType}"`) : '';
-  condition += '}, orderBy: id, orderDirection: asc';
-  const query = `
-    {
-      payments(${condition}) {
-        id
-        amount
-        rateToken {
-          address
-          decimals
-          name
-          symbol
-        }
-        paymentType
-        transactionHash
-        createdAt
-      }
-    }
-    `;
-  return query;
-};
-
 export const getFilteredServiceDescriptionCondition = (params: IProps): string => {
   let condition = ', where: {';
   condition += params.serviceStatus ? `service_: {status:"${params.serviceStatus}"}` : '';
